+++ internal/config/secrets.go
@@ -14,7 +14,7 @@ type Secrets struct {
 	SkipFilesAboveSize  int64             `yaml:"skip-files-above-size" json:"skip-files-above-size" mapstructure:"skip-files-above-size"`
 }
 
-func (cfg Secrets) loadDefaultConfig(v *viper.Viper) {
+func (cfg Secrets) loadDefaultValues(v *viper.Viper) {
 	v.SetDefault("secrets.cataloger.enabled", true)
 	v.SetDefault("secrets.cataloger.scope", source.AllLayersScope)
 	v.SetDefault("secrets.reveal-values", false)
